Silence conversion warnings

+++ groups/bal/ball/ball_fixedsizerecordbuffer.cpp
@@ -34,8 +34,8 @@ void FixedSizeRecordBuffer::popBack()
     d_currentTotalSize -=
         (d_deque.back())->numAllocatedBytes();
 
-    d_currentTotalSize -=
-           bsls::AlignmentUtil::roundUpToMaximalAlignment(sizeof(Record));
+    d_currentTotalSize -= static_cast<int>(
+               bsls::AlignmentUtil::roundUpToMaximalAlignment(sizeof(Record)));
 
     d_deque.pop_back();
 }
@@ -47,8 +47,8 @@ void FixedSizeRecordBuffer::popFront()
     d_currentTotalSize -=
         (d_deque.front())->numAllocatedBytes();
 
-    d_currentTotalSize -=
-           bsls::AlignmentUtil::roundUpToMaximalAlignment(sizeof(Record));
+    d_currentTotalSize -= static_cast<int>(
+               bsls::AlignmentUtil::roundUpToMaximalAlignment(sizeof(Record)));
 
     d_deque.pop_front();
 }
@@ -60,9 +60,11 @@ int FixedSizeRecordBuffer::pushBack(
 
 
     int size = handle->numAllocatedBytes() +
-           bsls::AlignmentUtil::roundUpToMaximalAlignment(sizeof(Record));
+           static_cast<int>(
+               bsls::AlignmentUtil::roundUpToMaximalAlignment(sizeof(Record)));
 
-    if (size + d_allocator.numBytesTotal() > d_maxTotalSize) {
+    if (size + static_cast<int>(d_allocator.numBytesTotal()) >
+                                                              d_maxTotalSize) {
         // impossible to accommodate this record
         return -1;                                                    // RETURN
     }
@@ -71,7 +73,8 @@ int FixedSizeRecordBuffer::pushBack(
     d_deque.push_back(handle); // This operation may cause 'd_deque' to grow,
                                // so test again.
 
-    if (size + d_allocator.numBytesTotal() > d_maxTotalSize) { // impossible
+    if (size + static_cast<int>(d_allocator.numBytesTotal()) >
+                                                d_maxTotalSize) { // impossible
                                                                // to
                                                                // accommodate
         d_deque.pop_back();
@@ -81,11 +84,13 @@ int FixedSizeRecordBuffer::pushBack(
         d_currentTotalSize += size;
     }
 
-    while(d_currentTotalSize + d_allocator.numBytesTotal() > d_maxTotalSize) {
+    while(d_currentTotalSize + static_cast<int>(d_allocator.numBytesTotal()) >
+                                                              d_maxTotalSize) {
         d_currentTotalSize -=
             (d_deque.front())->numAllocatedBytes();
         d_currentTotalSize -=
-           bsls::AlignmentUtil::roundUpToMaximalAlignment(sizeof(Record));
+            static_cast<int>(
+               bsls::AlignmentUtil::roundUpToMaximalAlignment(sizeof(Record)));
         d_deque.pop_front();
     }
     return returnValue;
@@ -98,8 +103,10 @@ int FixedSizeRecordBuffer::pushFront(
 
 
     int size = handle->numAllocatedBytes() +
-           bsls::AlignmentUtil::roundUpToMaximalAlignment(sizeof(Record));
-    if (size + d_allocator.numBytesTotal() > d_maxTotalSize) {
+           static_cast<int>(
+               bsls::AlignmentUtil::roundUpToMaximalAlignment(sizeof(Record)));
+    if (size + static_cast<int>(d_allocator.numBytesTotal()) >
+                                                              d_maxTotalSize) {
         // impossible to accommodate this record
         return -1;                                                    // RETURN
     }
@@ -108,7 +115,8 @@ int FixedSizeRecordBuffer::pushFront(
     d_deque.push_front(handle); // This operation may cause 'd_deque' to grow,
                                 // so test again.
 
-    if (size + d_allocator.numBytesTotal() > d_maxTotalSize) { // impossible
+    if (size + static_cast<int>(d_allocator.numBytesTotal()) >
+                                                d_maxTotalSize) { // impossible
                                                                // to
                                                                // accommodate
         d_deque.pop_front();
@@ -118,11 +126,13 @@ int FixedSizeRecordBuffer::pushFront(
         d_currentTotalSize += size;
     }
 
-    while(d_currentTotalSize + d_allocator.numBytesTotal() > d_maxTotalSize) {
+    while(d_currentTotalSize + static_cast<int>(d_allocator.numBytesTotal()) >
+                                                              d_maxTotalSize) {
         d_currentTotalSize -=
-            (d_deque.back())->numAllocatedBytes();
+            static_cast<int>((d_deque.back())->numAllocatedBytes());
         d_currentTotalSize -=
-           bsls::AlignmentUtil::roundUpToMaximalAlignment(sizeof(Record));
+            static_cast<int>(
+               bsls::AlignmentUtil::roundUpToMaximalAlignment(sizeof(Record)));
         d_deque.pop_back();
     }
     return returnValue;
